sign up page done

+++ app/javascript/editor/editor_show_button.js
@@ -12,16 +12,18 @@ document.addEventListener('turbolinks:load', () => {
   // let penDelParams = `user[username]=${userName}&pen[random_url]=${randomURL}`
   let penDelParams = `pen[random_url]=${randomURL}`
 
-  delBtn.addEventListener('click', () => {
-    Rails.ajax({
-      url: editPagePath,
-      type: 'DELETE',
-      headers: {
-        "Accept": "application/json",
-        "Content-Type": "application/json"
-      },
-      dataType: 'json',
-      data: penDelParams
+  if (delBtn) {
+    delBtn.addEventListener('click', () => {
+      Rails.ajax({
+        url: editPagePath,
+        type: 'DELETE',
+        headers: {
+          "Accept": "application/json",
+          "Content-Type": "application/json"
+        },
+        dataType: 'json',
+        data: penDelParams
+      })
     })
-  })
+  }
 })
